Support starting unicorn master process as non-root, usable if using unveil instead of chroot
If you are using unveil, there is no need to start the process as
root.  root access was only required for chroot.  The priv dropping
was useful for limiting file system access as well inside the
application, but that can be done with unveil when not dropping
privileges.

To get this to work, use a separate directory for the files that
hold request specific information.  Instead of only being accessable
to root, this allows execute to the _unicorn group.  Add per-app
subdirectories that limit access to the application-specific user
to hold the files.

Don't try to drop privileges if not running as root when sending
error emails.

Add a -N option to unicorn-lockdown-add to set the daemon_user
inside the rc.d file so that the unicorn master process will not
start as root.