Remove support for chroot and starting applications as root
This allows us to remove a lot of code that previously had to
handle chrooting and privilege dropping.

Since starting the application as root is no longer an option,
we don't need to call Unicorn::Worker#user to change permissions,
so we can drop the :user and :group options to Unicorn.lockdown.