Perform application tests in CI
To actually test unicorn-lockdown requires root access, both to create
the necessary files and directories, and to manage the unicorn
processes with rcctl.

We don't usually want to run most tests as root.  However, since CI
should run the tests as root anyway, and we don't need to worry about
making changes to the local environment during CI, since a new VM is
created for each CI run, it seems like a good idea to actual test for
correct behavior in CI.

This adds a test suite that runs unicorn-lockdown-setup and
unicorn-lockdown-add, adjusts the nginx configuration, and adds
a config.ru file, then starts up nginx and the unicorn process.  Then
it tests that requests work, that the unveil and pledge protections
work as expected, and that rcctl {start,stop,reload,check} works as
expected.